+++ app/models/PayPeriodDetail.scala
@@ -19,7 +19,7 @@ package models
 import play.api.libs.json.{Json, OFormat, Reads, Writes, __}
 
 case class PayPeriodDetail(
-                            amount:       Int,
+                            amount:       BigDecimal,
                             howManyAWeek: BigDecimal,
                             period:       String,
                             urlForChange: String)
@@ -28,7 +28,7 @@ object PayPeriodDetail {
   import play.api.libs.functional.syntax._
 
   implicit lazy val reads: Reads[PayPeriodDetail] = (
-    (__ \ "amount").read[Int] and
+    (__ \ "amount").read[BigDecimal] and
       (__ \ "how-many-a-week").read[BigDecimal] and
         (__ \ "period").read[String] and
         (__ \ "urlForChange").read[String]
@@ -37,7 +37,7 @@ object PayPeriodDetail {
 
   implicit lazy val writes: Writes[PayPeriodDetail] =
     (
-      (__ \ "amount").write[Int] and
+      (__ \ "amount").write[BigDecimal] and
         (__ \ "how-many-a-week").write[BigDecimal] and
           (__ \ "period").write[String] and
         (__ \ "urlForChange").write[String]
